requirements.txt
absl-py==0.4.1
astor==0.7.1
atari-py==0.1.1
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
future==0.16.0
gast==0.2.0
grpcio==1.15.0
gym==0.10.5
h5py==2.8.0
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==2.6.11
numpy==1.14.5
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
progressbar2==3.38.0
prompt-toolkit==2.0.6
protobuf==3.6.1
ptyprocess==0.6.0
pyglet==1.3.2
Pygments==2.2.0
PyOpenGL==3.1.0
python-utils==2.3.0
PyYAML==3.13
req==1.0.0
requests==2.19.1
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
torch==0.4.1
torchvision==0.2.1
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
